Make secondaries aware of idempotency ID in append
In this manner, the secondaries will have the status of the idempotency
ID within their memory. So, if primary does down, we will have
information of append completion even within the secondaries. If primary
as well as all secondaries go down, idempotency will be lost but we
assume that this does not happen.